requirements.txt
fastapi
uvicorn
gitpython
scikit-learn
pandas
docker
joblib
api
python-multipart
xgboost
gensim
accelerate

# CPU-only PyTorch + related libs
torch==2.3.1+cpu
torchvision==0.18.1+cpu
torchaudio==2.3.1+cpu
transformers==4.44.2
-f https://download.pytorch.org/whl/cpu/torch_stable.html
